+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 ### (TBA)
 - Fix: Profile Image upload/edit
+- Enh: VerifiedIcon widget options
 
 ### 1.0.0-beta.3 (4/20/2023)
 - Enh: People and spaces page